// web/src/app/api/model/authored-module-response.ts
/**
 * Simplicity training API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { VersionResponse } from './version-response';


export interface AuthoredModuleResponse { 
    assignedTeamIds?: Array<string>;
    createdAt?: string;
    draft?: VersionResponse;
    moduleId?: string;
    published?: VersionResponse;
    summary?: string;
    title?: string;
}