// web/src/app/api/model/module-summary-response.ts
/**
 * Simplicity training API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ModuleSummaryResponse { 
    assignedTeamCount?: number;
    hasDraft?: boolean;
    moduleId?: string;
    publishedVersion?: number;
    summary?: string;
    title?: string;
}