// web/src/app/api/model/question-input.ts
/**
 * Simplicity training API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OptionInput } from './option-input';


export interface QuestionInput { 
    explanation?: string;
    options: Array<OptionInput>;
    prompt: string;
}