// web/src/app/api/model/section-input.ts
/**
 * Simplicity training API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SectionInput { 
    body?: string;
    mediaAssetId?: string;
    title: string;
}