// web/src/app/api/model/team-response.ts
/**
 * Simplicity training API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TeamResponse { 
    createdAt?: string;
    description?: string;
    id?: string;
    memberCount?: number;
    name?: string;
    orgId?: string;
}